task

+++ cyberlarp-api/src/main/java/com/example/cyberlarpapi/game/controllers/CharacterController.java
@@ -181,7 +181,7 @@ public class CharacterController {
         private Integer knowledge;
         private Integer maxHp;
         private Integer currentHp;
-        private Integer balance;
+        private Float balance;
     }
 
     @Getter
@@ -223,7 +223,7 @@ public class CharacterController {
             private Integer knowledge;
             private Integer maxHp;
             private Integer currentHp;
-            private Integer balance;
+            private Float balance;
             private String accountNumber;
             private Integer armor;
